feat: Add promise.allSettled examples for handling multiple promises
The code changes in `promise-allSettled/index.js` and `promise-allSettled/index2.js` demonstrate the usage of `Promise.allSettled` to handle multiple promises.

In `promise-allSettled/index.js`, two promises `p1` and `p2` are created using `Promise.resolve` and `new Promise` respectively. `Promise.allSettled` is then used to wait for both promises to settle. The results are logged to the console.

In `promise-allSettled/index2.js`, an array of promises is created using `Promise.resolve`, `new Promise`, and a non-promise value. `Promise.allSettled` is used to wait for all promises to settle. The settled values are logged to the console.

These code changes provide examples of how to use `Promise.allSettled` to handle multiple promises and handle both fulfilled and rejected promises.

Note: The recent user commits and repository commits are not relevant to the suggested commit message.