feat: Add Promise.race example for handling multiple promises
The code changes in `topics/async-js/promise/promise-concurrency/race/index.js` demonstrate the usage of `Promise.race` to handle multiple promises. Two promises, `promise1` and `promise2`, are created with different timeouts. The `Promise.race` method is then used to wait for the first promise to settle and log its value to the console.

This code change provides an example of how to use `Promise.race` to handle multiple promises and retrieve the result of the fastest settling promise.

Note: The recent user commits and repository commits are not relevant to the suggested commit message.